Update to test and added nickname proportion.  TODO get scale to work
