+++ src/pseudopeople/constants/metadata.py
@@ -23,3 +23,9 @@ DATEFORMATS = __DateFormats()
 
 class Attributes:
     DATE_FORMAT = "date_format"
+
+
+# Value calculated for noise scaling for nicknames
+# Constant calculated by number of names with nicknames / number of names used in PRL name mapping
+# Found in PRL data.nickname_proportions.get_nickname_proportion()
+NICKNAMES_PROPORTION = 0.033635429318882554
