+++ src/pseudopeople/noise_entities.py
@@ -1,6 +1,6 @@
 from typing import NamedTuple
 
-from pseudopeople import noise_functions, utilities
+from pseudopeople import noise_functions, utilities, noise_scaling
 from pseudopeople.configuration import Keys
 from pseudopeople.entity_types import ColumnNoiseType, RowNoiseType
 
@@ -27,7 +27,7 @@ class __NoiseTypes(NamedTuple):
     incorrect_selection: ColumnNoiseType = ColumnNoiseType(
         "choose_wrong_option",
         noise_functions.generate_incorrect_selections,
-        noise_level_scaling_function=utilities.noise_scaling_incorrect_selection,
+        noise_level_scaling_function=noise_scaling.noise_scaling_incorrect_selection,
     )
     # copy_from_within_household: ColumnNoiseType = ColumnNoiseType(
     #     "copy_from_household_member",
@@ -61,7 +61,7 @@ class __NoiseTypes(NamedTuple):
     nickname: ColumnNoiseType = ColumnNoiseType(
         "use_nickname",
         noise_functions.generate_nicknames,
-        noise_level_scaling_function=utilities.scale_nicknames,
+        noise_level_scaling_function=noise_scaling.scale_nicknames,
     )
     fake_name: ColumnNoiseType = ColumnNoiseType(
         "use_fake_name",
